requirements.txt
email-validator>=2.0.0
dnspython>=2.3.0
requests>=2.31.0
